added the method needed

+++ print_random.c
@@ -1,19 +1,18 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
-
-
+#include "random.c"
 
 int main()
 {
 	int a;
 
 	/* seed the randomizer */
-	srand( (unsigned)time(NULL) );
+    srand(time(NULL));
 
-	printf("Today's random word: ");
+    printf("Today's random word: ");
 	for(a=0;a<7;a++)
-		putchar( randchar() );
+		putchar(randchar());
 	putchar('\n');
 
 	return(0);
